Migrate explode code to shared
